Create cachedir before volume mounting it
When volume mounting a dir that does not exist on the host machine docker will normally create that dir.
It will also do this on macos using Docker for mac, except it does this inside the linux vm that is used to actually run docker.
Therefor we end up with the wring cache dir on mac if we rely on docker creating it for us. So lets do it ourselves.