Tolerate needs.pipeline (#732)
When a job has needs.pipeline, because it depends in some way on a job
from another pipeline we currently fail because the needed job isn't in
the pipeline's jobs. This means that we cannot test a pipeline that is
part of a multi-project or parent-child pipeline configuration, even if
the dependency can be mocked somehow (e.g. passing some variable on the
CLI). Instead, it is more useful to ignore the particular dependency and
emit a warning.

I made a choice to include the pipeline attribute in the Job.needs data
structure, which would indicate to the validator and executor that the
specific dependency should be ignored. I also updated the Validator to
generate a list of warnings that can be output all together.
Alternatives I considered were a) updating the Job.needs method to just
skip the entries of Job.needs that contained the pipeline field because
that method isn't memoized so I couldn't emit a warning there (if you
would prefer we memoize it then this alternative would be a good one),
and b) passing the writeStreams object to the Validator so that it could
output the warnings itself, but I feel the parser might be a better
place to control the "interaction" with the user.

Former-commit-id: 42fe571443a73b7c01a96bcb5aea09b04dc7c4b3