Trivial: Cast dockerCmd string replace to string.
Issue: with one of my gitlab pipelines, our usage of
service directives over parallel executors was causing node errors
to be raised when trying to replace dockerCmd strings that
for some reason were not strings.

Solution: Force casting the values to be string representations
in all cases fixed the issue I was having and is a safer
call to make.